+++ .gitignore
@@ -1,3 +1,5 @@
 build/bin
 node_modules
 frontend/dist
+frontend/.yarn
+frontend/.yarnrc.yml
