+++ frontend/package.json.md5
@@ -1,1 +1,1 @@
-3680670718fd2e138094436e6bad6da4
+148954b5c5691fa9347ca754d5e1d760
