+++ frontend/src/components/DataTable/components/TableHeader.tsx
@@ -1,0 +1,56 @@
+import {
+    createStyles,
+    Group,
+    UnstyledButton,
+    Text,
+    Center,
+} from '@mantine/core';
+import { IconChevronDown, IconChevronUp, IconSelector } from '@tabler/icons';
+import { ThProps } from '../../../shared/interfaces/th-props.interface';
+
+const useStyles = createStyles((theme) => ({
+    th: {
+        padding: '0 !important',
+    },
+
+    control: {
+        width: '100%',
+        padding: `${theme.spacing.xs}px ${theme.spacing.md}px`,
+
+        '&:hover': {
+            backgroundColor:
+                theme.colorScheme === 'dark'
+                    ? theme.colors.dark[6]
+                    : theme.colors.gray[0],
+        },
+    },
+
+    icon: {
+        width: 21,
+        height: 21,
+        borderRadius: 21,
+    },
+}));
+
+export function TableHeader({ children, reversed, sorted, onSort }: ThProps) {
+    const { classes } = useStyles();
+    const Icon = sorted
+        ? reversed
+            ? IconChevronUp
+            : IconChevronDown
+        : IconSelector;
+    return (
+        <th className={classes.th}>
+            <UnstyledButton onClick={onSort} className={classes.control}>
+                <Group position="apart">
+                    <Text weight={500} size="sm">
+                        {children}
+                    </Text>
+                    <Center className={classes.icon}>
+                        <Icon size={14} stroke={1.5} />
+                    </Center>
+                </Group>
+            </UnstyledButton>
+        </th>
+    );
+}
