+++ frontend/src/components/NavSide/NavSide.tsx
@@ -1,228 +1,7 @@
-import { useState } from "react";
-import {
-  Center,
-  Badge,
-  Avatar,
-  Navbar,
-  SegmentedControl,
-  Text,
-  createStyles,
-} from "@mantine/core";
-
-import {
-  IconShoppingCart,
-  IconLicense,
-  IconMessage2,
-  IconBellRinging,
-  IconMessages,
-  IconFingerprint,
-  IconKey,
-  IconSettings,
-  Icon2fa,
-  IconUsers,
-  IconFileAnalytics,
-  IconDatabaseImport,
-  IconReceipt2,
-  IconReceiptRefund,
-  IconLogout,
-  IconSwitchHorizontal,
-  IconLock,
-  IconNotes,
-  IconCloudLock,
-  IconLockAccess,
-  IconFilePower,
-} from "@tabler/icons";
-import { useNavigate } from "react-router-dom";
-
-const useStyles = createStyles((theme, _params, getRef) => {
-  const icon = getRef("icon");
-
-  return {
-    navbar: {
-      backgroundColor:
-        theme.colorScheme === "dark" ? theme.colors.dark[7] : theme.white,
-    },
-
-    title: {
-      fontSize: "1.2em",
-      textTransform: "lowercase",
-      letterSpacing: -0.25,
-    },
-
-    link: {
-      ...theme.fn.focusStyles(),
-      display: "flex",
-      alignItems: "center",
-      textDecoration: "none",
-      fontSize: theme.fontSizes.sm,
-      color:
-        theme.colorScheme === "dark"
-          ? theme.colors.dark[1]
-          : theme.colors.gray[7],
-      padding: `${theme.spacing.xs}px ${theme.spacing.sm}px`,
-      borderRadius: theme.radius.sm,
-      fontWeight: 500,
-
-      "&:hover": {
-        backgroundColor:
-          theme.colorScheme === "dark"
-            ? theme.colors.dark[6]
-            : theme.colors.gray[0],
-        color: theme.colorScheme === "dark" ? theme.white : theme.black,
-
-        [`& .${icon}`]: {
-          color: theme.colorScheme === "dark" ? theme.white : theme.black,
-        },
-      },
-    },
-
-    linkIcon: {
-      ref: icon,
-      color:
-        theme.colorScheme === "dark"
-          ? theme.colors.dark[2]
-          : theme.colors.gray[6],
-      marginRight: theme.spacing.sm,
-    },
-
-    linkActive: {
-      "&, &:hover": {
-        backgroundColor: theme.fn.variant({
-          variant: "light",
-          color: theme.primaryColor,
-        }).background,
-        color: theme.fn.variant({ variant: "light", color: theme.primaryColor })
-          .color,
-        [`& .${icon}`]: {
-          color: theme.fn.variant({
-            variant: "light",
-            color: theme.primaryColor,
-          }).color,
-        },
-      },
-    },
-
-    footer: {
-      borderTop: `1px solid ${theme.colorScheme === "dark"
-        ? theme.colors.dark[4]
-        : theme.colors.gray[3]
-        }`,
-      paddingTop: theme.spacing.md,
-    },
-  };
-});
-
-const tabs = {
-  account: [
-    { link: "/notifications", label: "Notifications", icon: IconBellRinging },
-    { link: "/notes", label: "Notes", icon: IconNotes },
-    { link: "/password_manager", label: "Password Manager", icon: IconLockAccess },
-    { link: "/password_generator", label: "Password Generator", icon: IconKey },
-    { link: "/encrypt_files", label: "Encrypt Files", icon: IconFilePower },
-    { link: "/settings", label: "Settings", icon: IconSettings },
-  ],
-  general: [
-    { link: "/", label: "Databases", icon: IconDatabaseImport },
-    { link: "", label: "Orders", icon: IconShoppingCart },
-    { link: "", label: "Receipts", icon: IconLicense },
-    { link: "", label: "Reviews", icon: IconMessage2 },
-    { link: "", label: "Messages", icon: IconMessages },
-    { link: "", label: "Customers", icon: IconUsers },
-    { link: "", label: "Refunds", icon: IconReceiptRefund },
-    { link: "", label: "Files", icon: IconFileAnalytics },
-  ],
-};
-
-function BadgeWithAvatar({ title }: { title: string }) {
-  return (
-    <>
-      <Avatar
-        alt="Avatar for badge"
-        size={50}
-        radius={25}
-        mr={5}
-        src="https://i.pravatar.cc/150?u=edrx"
-      />
-      {/* Will Come back to this later */}
-      <Badge sx={{ textTransform: "none" }} size="xl" radius="xl" color="teal">
-        {title}
-      </Badge>
-    </>
-  );
-}
-
-function NavbarSegmented() {
-  const navigate = useNavigate();
-  const { classes, cx } = useStyles();
-  const [section, setSection] = useState<"account" | "general">("account");
-  const [active, setActive] = useState("Billing");
-
-  const links = tabs[section].map((item) => (
-    <a
-      className={cx(classes.link, {
-        [classes.linkActive]: item.label === active,
-      })}
-      href={item.link}
-      key={item.label}
-      onClick={(event) => {
-        event.preventDefault();
-        setActive(item.label);
-        navigate(item.link);
-      }}
-    >
-      <item.icon className={classes.linkIcon} stroke={1.5} />
-      <span>{item.label}</span>
-    </a>
-  ));
-
-  return (
-    <Navbar width={{ sm: 300 }} p="md" className={classes.navbar}>
-      <Navbar.Section>
-        <Center sx={{ marginBottom: "0.7rem" }}>
-          <BadgeWithAvatar title={"user@example.com"} />
-        </Center>
-
-        <SegmentedControl
-          value={section}
-          onChange={(value: "account" | "general") => setSection(value)}
-          transitionTimingFunction="ease"
-          fullWidth
-          data={[
-            { label: "Account", value: "account" },
-            { label: "System", value: "general" },
-          ]}
-        />
-      </Navbar.Section>
-
-      <Navbar.Section grow mt="xl">
-        {links}
-      </Navbar.Section>
-
-      <Navbar.Section className={classes.footer}>
-        <a
-          href="#"
-          className={classes.link}
-          onClick={(event) => event.preventDefault()}
-        >
-          <IconSwitchHorizontal className={classes.linkIcon} stroke={1.5} />
-          <span>Change account</span>
-        </a>
-
-        <a
-          href="#"
-          className={classes.link}
-          onClick={(event) => event.preventDefault()}
-        >
-          <IconLogout className={classes.linkIcon} stroke={1.5} />
-          <span>Logout</span>
-        </a>
-      </Navbar.Section>
-    </Navbar>
-  );
-}
+import { NavbarSegmented } from "./components/NavbarSegmented";
 
 function NavSide() {
-  return <NavbarSegmented />;
+	return <NavbarSegmented />;
 }
 
 export default NavSide;
