--- frontend/src/pages/PasswordGenerator/PasswordGenerator.module.css
@@ -1,6 +1,0 @@
-.container {
-    height: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
