+++ frontend/src/shared/interfaces/row-data.interface.ts
@@ -1,0 +1,3 @@
+export interface RowData {
+    [key: string]: string;
+}
