+++ frontend/src/shared/interfaces/table-sort-props.interface.ts
@@ -1,0 +1,5 @@
+import { RowData } from "./row-data.interface";
+
+export interface TableSortProps {
+	data: RowData[];
+}
