+++ frontend/src/shared/interfaces/th-props.interface.ts
@@ -1,0 +1,6 @@
+export interface ThProps {
+	children: React.ReactNode;
+	reversed: boolean;
+	sorted: boolean;
+	onSort(): void;
+}
